Move handshake & send/recv common code to relay.rs (rename relay.rs -> send_recv.rs)
