+++ iroh-relay/src/protos.rs
@@ -3,5 +3,6 @@
 pub mod disco;
 pub mod handshake;
 pub mod relay;
+pub mod send_recv;
 pub mod streams;
 pub mod stun;
