+++ iroh-relay/src/server.rs
@@ -894,7 +894,7 @@ mod tests {
         dns::DnsResolver,
         protos::{
             self,
-            relay::{ClientToServerMsg, ServerToClientMsg},
+            send_recv::{ClientToServerMsg, ServerToClientMsg},
         },
     };
